requirements-dev.txt
django-debug-toolbar==3.2.4
pylint==2.13.7
codecov==2.1.12
